update conv seq2seq.
