+++ examples/t5/README.md
@@ -55,7 +55,7 @@ python predict.py
 
 
 
-## Dataset
+## ContextDataset
 
 | 数据集 | 语料 | 下载链接 | 压缩包大小 |
 | :------- | :--------- | :---------: | :---------: |
