--- pycorrector/deepcontext/config.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-@author:XuMing(xuming624@qq.com)
-@description: network configuration
-"""
-
