+++ pycorrector/gpt/merge_peft_adapter.py
@@ -67,10 +67,10 @@ def main():
     if args.tokenizer_path:
         tokenizer = tokenizer_class.from_pretrained(args.tokenizer_path, trust_remote_code=True)
     else:
-        tokenizer = tokenizer_class.from_pretrained(peft_model_path, trust_remote_code=True)
+        tokenizer = tokenizer_class.from_pretrained(base_model_path, trust_remote_code=True)
     if args.resize_emb:
         base_model_token_size = base_model.get_input_embeddings().weight.size(0)
-        if base_model_token_size != len(tokenizer):
+        if base_model_token_size < len(tokenizer):
             base_model.resize_token_embeddings(len(tokenizer))
             print(f"Resize vocabulary size {base_model_token_size} to {len(tokenizer)}")
 
@@ -86,7 +86,7 @@ def main():
 
     print("Saving to Hugging Face format...")
     tokenizer.save_pretrained(output_dir)
-    base_model.save_pretrained(output_dir, max_shard_size='10GB', safe_serialization=True)
+    base_model.save_pretrained(output_dir, max_shard_size='10GB', safe_serialization=False)
     print(f"Done! model saved to {output_dir}")
