+++ pycorrector/seq2seq/conv_seq2seq_corrector.py
@@ -9,15 +9,14 @@ import sys
 import time
 from typing import List
 
-import numpy as np
 import torch
 from loguru import logger
 
 sys.path.append('../..')
 
-from pycorrector.seq2seq.data_reader import SOS_TOKEN, EOS_TOKEN, PAD_TOKEN, load_word_dict
-from pycorrector.seq2seq.conv_seq2seq import ConvSeq2Seq
+from pycorrector.seq2seq.conv_seq2seq_model import ConvSeq2SeqModel
 
+from pycorrector.utils.tokenizer import split_text_into_sentences_by_length
 from pycorrector.utils.get_file import get_file
 from pycorrector.detector import USER_DATA_DIR
 
@@ -30,13 +29,10 @@ pre_trained_seq2seq_models = {
 }
 
 
-class Seq2SeqCorrector:
+class ConvSeq2SeqCorrector:
     def __init__(
             self,
             model_name_or_path: str = '',
-            embed_size: int = 128,
-            hidden_size: int = 128,
-            dropout: float = 0.25,
             max_length: int = 128,
     ):
         bin_path = os.path.join(model_name_or_path, 'convseq2seq.pth')
@@ -57,28 +53,9 @@ class Seq2SeqCorrector:
             )
         t1 = time.time()
         logger.debug("Device: {}".format(device))
-        src_vocab_path = os.path.join(model_dir, 'vocab_source.txt')
-        trg_vocab_path = os.path.join(model_dir, 'vocab_target.txt')
-        self.src_2_ids = load_word_dict(src_vocab_path)
-        self.trg_2_ids = load_word_dict(trg_vocab_path)
-        self.id_2_trgs = {v: k for k, v in self.trg_2_ids.items()}
-        trg_pad_idx = self.trg_2_ids[PAD_TOKEN]
-        self.model = ConvSeq2Seq(
-            encoder_vocab_size=len(self.src_2_ids),
-            decoder_vocab_size=len(self.trg_2_ids),
-            embed_size=embed_size,
-            enc_hidden_size=hidden_size,
-            dec_hidden_size=hidden_size,
-            dropout=dropout,
-            trg_pad_idx=trg_pad_idx,
-            device=device,
-            max_length=max_length
-        ).to(device)
-        model_path = os.path.join(model_dir, 'convseq2seq.pth')
-        self.model.load_state_dict(torch.load(model_path, map_location=device))
-        self.model.eval()
+        self.model = ConvSeq2SeqModel(model_dir=model_dir)
         self.max_length = max_length
-        logger.debug('Loaded model: %s, spend: %.4f s.' % (model_path, time.time() - t1))
+        logger.debug('Loaded model: %s, spend: %.4f s.' % (model_dir, time.time() - t1))
 
     @staticmethod
     def get_errors(corrected_text, origin_text):
@@ -95,46 +72,48 @@ class Seq2SeqCorrector:
         sub_details = sorted(sub_details, key=operator.itemgetter(2))
         return corrected_text, sub_details
 
-    def predict(self, sentences: List[str]):
-        corrected_sents = []
-        details = []
-        for query in sentences:
-            out = []
-            tokens = [token.lower() for token in query]
-            tokens = [SOS_TOKEN] + tokens + [EOS_TOKEN]
-            src_ids = [self.src_2_ids[i] for i in tokens if i in self.src_2_ids]
-
-            sos_idx = self.trg_2_ids[SOS_TOKEN]
-            src_tensor = torch.from_numpy(np.array(src_ids).reshape(1, -1)).long().to(device)
-            translation, attn = self.model.translate(src_tensor, sos_idx)
-            translation = [self.id_2_trgs[i] for i in translation if i in self.id_2_trgs]
-            for word in translation:
-                if word != EOS_TOKEN:
-                    out.append(word)
-                else:
-                    break
-            corrected_sent = ''.join(out)
-            corrected_sent, sub_details = self.get_errors(corrected_sent, query)
-            corrected_sents.append(corrected_sent)
-            details.append(sub_details)
-        return corrected_sents, details
-
-    def correct_batch(self, sentences: List[str]):
+    def correct_batch(self, sentences: List[str], max_length: int = 128, silent: bool = True):
         """
         批量句子纠错
         :param: sentences, List[str]: 待纠错的句子
+        :param: max_length, int: 句子最大长度
+        :param: silent, bool: 是否打印日志
         :return: list, [corrected_texts, [error_word, correct_word, begin_pos, end_pos]]
         """
-        return self.predict(sentences)
+        input_sents = []
+        sent_map = []
+        for idx, sentence in enumerate(sentences):
+            if len(sentence) > max_length:
+                # split long sentence into short ones
+                short_sentences = [i[0] for i in split_text_into_sentences_by_length(sentence, max_length)]
+                input_sents.extend(short_sentences)
+                sent_map.extend([idx] * len(short_sentences))
+            else:
+                input_sents.append(sentence)
+                sent_map.append(idx)
 
-    def correct(self, sentence: str, **kwargs):
+        # batch predict
+        corrected_sents = self.model.predict(input_sents, silent=silent)
+
+        # concatenate the results of short sentences
+        corrected_sentences = [''] * len(sentences)
+        for idx, corrected_sent in zip(sent_map, corrected_sents):
+            corrected_sentences[idx] += corrected_sent
+        corrected_details = []
+        for idx, corrected_sent in enumerate(corrected_sentences):
+            _, sub_details = self.get_errors(corrected_sent, sentences[idx])
+            corrected_details.append(sub_details)
+
+        return corrected_sentences, corrected_details
+
+    def correct(self, sentence: str):
         """Correct a sentence with conv seq2seq model"""
         corrected_sentences, corrected_details = self.correct_batch([sentence])
         return corrected_sentences[0], corrected_details[0]
 
 
 if __name__ == "__main__":
-    m = Seq2SeqCorrector()
+    m = ConvSeq2SeqCorrector()
     error_sentences = [
         '老是较书。',
         '感谢等五分以后，碰到一位很棒的奴生跟我可聊。',
@@ -143,6 +122,6 @@ if __name__ == "__main__":
         '他们只能有两个选择：接受降新或自动离职。',
         '王天华开心得一直说话。'
     ]
-    res, ds = m.correct_batch(error_sentences)
+    res, ds = m.correct_batch(error_sentences, silent=False)
     for sent, r, d in zip(error_sentences, res, ds):
         print("original sentence:{} => {} , err:{}".format(sent, r, d))
